Django-Middleware-0x03/0x03-MessagingApp-Django/pyproject.toml
[tool.poetry]
name = "0x03-messagingapp-django"
version = "0.1.0"
description = ""
authors = ["ForeverCountry <40751841+ForeverCountry@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
django-rest-framework-nested = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
